release/requirements.txt
GitPython==3.1.44
pygithub==2.5.0
